requirements.txt
flask
google-adk
google-genai
python-dotenv
waitress
